+++ tests/test_main.py
@@ -1024,6 +1024,98 @@ class TestInit(unittest.TestCase):
         shutil.rmtree(output_directory)
         self.assertFalse(os.path.isdir(output_directory))
 
+    def test_init_move_input_files(self):
+        """
+        Test the init code with valid parameters, multiple input files
+        """
+        input_directory = "tests/sample_files/multiple"
+        output_directory = "nonexistent_dir"
+        self.assertEqual(num_files_in_directory(input_directory), 3)
+        self.assertFalse(os.path.isdir(output_directory))
+        target.__name__ = "__main__"
+        target.sys.argv = [
+            "scriptname.py",
+            "--input-directory",
+            input_directory,
+            "--output-directory",
+            output_directory,
+            "--move-input-files",
+            "--overwrite-files",
+            "--config",
+            "tests/sample_files/configuration1.xlsx",
+            "--delimiter",
+            "^",
+            "--skip-header",
+            "1",
+            "--skip-footer",
+            "1",
+            "--application-id",
+            "SE",
+            "--run-description",
+            "AAA",
+            "--billing-type",
+            "H",
+            "--run-id",
+            "123",
+            "--txt-extension",
+            "--locale",
+            "fr_FR.utf8",
+        ]
+        try:
+            target.init()
+        except localeError:
+            # On Windows, the French locale is just called "fr"
+            target.sys.argv[-1] = "fr"
+            target.init()
+
+        self.assertTrue(os.path.isdir(output_directory))
+        # The 3 input files and the 6 generated output files
+        self.assertEqual(num_files_in_directory(output_directory), 3 + 6)
+
+        for i in ("", "_copy1", "_copy2"):
+            # Move the input files back to the input directory
+            input_file = os.path.join(output_directory, "input1%s.txt" % i)
+            shutil.move(input_file, input_directory)
+
+        for run_id in (123, 124, 125):
+            # Confirm the output files have been written and their content
+            metadata_file_name = "%s/SSE0%dE.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(metadata_file_name))
+            with open(metadata_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "SSEAAA                           "
+                    "9999999900000000H00000300%dV1.11                          "
+                    "                                                           "
+                    "                                                 " % run_id
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the metadata output file
+            os.remove(metadata_file_name)
+            self.assertFalse(os.path.isfile(metadata_file_name))
+
+            detailed_file_name = "%s/SSE0%dD.txt" % (output_directory, run_id)
+            self.assertTrue(os.path.isfile(detailed_file_name))
+            with open(detailed_file_name) as f:
+                s = f.read()
+                expected_output = (
+                    "0004000133034205413540000100202007312006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034005407940000157202003051022"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           \n"
+                    "0004000133034105409340022139202012252006"
+                    "                                        "
+                    "Leendert MOLENDIJK [90038979]           "
+                )
+                self.assertEqual(expected_output, s)
+            # Remove the detailed output file
+            os.remove(detailed_file_name)
+            self.assertFalse(os.path.isfile(detailed_file_name))
+        shutil.rmtree(output_directory)
+        self.assertFalse(os.path.isdir(output_directory))
+
 
 class TestLicense(unittest.TestCase):
     def test_license_file(self):